tutors_assistant/pyproject.toml
[tool.poetry]
name = "tutors assistant"
version = "0.1.0"
description = ""
authors = ["garmoshka"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
pytest = "^7.1.1"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
psycopg2 = "^2.9.3"
black = "^22.3.0"
pytest-cov = "^3.0.0"
psycopg2-binary = "^2.9.3"
Pillow = "^9.0.1"
django-singleton-model = "^0.2.3"
aiogram = "^2.19"
validate_email = "^1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
